+++ core/state/state_frozen_fund.go
@@ -127,7 +127,7 @@ func (c *stateFrozenFund) punishFund(context *StateDB, candidateAddress [20]byte
 			})
 
 			item.Value = newValue
-			context.DeleteCoinIfZeroReserve(item.Coin)
+			context.SanitizeCoin(item.Coin)
 		}
 
 		newList[i] = item