+++ core/transaction/sell_all_coin.go
@@ -205,8 +205,8 @@ func (data SellAllCoinData) Run(tx *Transaction, context *state.StateDB, isCheck
 		context.AddBalance(sender, data.CoinToBuy, value)
 		context.SetNonce(sender, tx.Nonce)
 
-		context.DeleteCoinIfZeroReserve(data.CoinToBuy)
-		context.DeleteCoinIfZeroReserve(data.CoinToSell)
+		context.SanitizeCoin(data.CoinToBuy)
+		context.SanitizeCoin(data.CoinToSell)
 	}
 
 	tags := common.KVPairs{